+++ src/Symfony/Component/Serializer/Annotation/MaxDepth.php
@@ -21,6 +21,7 @@ use Symfony\Component\Serializer\Exception\InvalidArgumentException;
  *
  * @author Kévin Dunglas <dunglas@gmail.com>
  */
+#[\Attribute(\Attribute::TARGET_METHOD | \Attribute::TARGET_PROPERTY)]
 class MaxDepth
 {
     /**
@@ -28,17 +29,23 @@ class MaxDepth
      */
     private $maxDepth;
 
-    public function __construct(array $data)
+    /**
+     * @param int|array $maxDepth
+     */
+    public function __construct($maxDepth)
     {
-        if (!isset($data['value'])) {
-            throw new InvalidArgumentException(sprintf('Parameter of annotation "%s" should be set.', static::class));
+        if (\is_array($maxDepth)) {
+            if (!isset($maxDepth['value'])) {
+                throw new InvalidArgumentException(sprintf('Parameter of annotation "%s" should be set.', static::class));
+            }
+            $maxDepth = $maxDepth['value'];
         }
 
-        if (!\is_int($data['value']) || $data['value'] <= 0) {
+        if (!\is_int($maxDepth) || $maxDepth <= 0) {
             throw new InvalidArgumentException(sprintf('Parameter of annotation "%s" must be a positive integer.', static::class));
         }
 
-        $this->maxDepth = $data['value'];
+        $this->maxDepth = $maxDepth;
     }
 
     public function getMaxDepth()