+++ src/Symfony/Component/Serializer/Tests/Fixtures/Annotations/AbstractDummyThirdChild.php
@@ -9,7 +9,7 @@
  * file that was distributed with this source code.
  */
 
-namespace Symfony\Component\Serializer\Tests\Fixtures;
+namespace Symfony\Component\Serializer\Tests\Fixtures\Annotations;
 
 final class AbstractDummyThirdChild extends AbstractDummyFirstChild
 {