+++ src/Symfony/Component/Serializer/Tests/Fixtures/Annotations/GroupDummy.php
@@ -9,9 +9,10 @@
  * file that was distributed with this source code.
  */
 
-namespace Symfony\Component\Serializer\Tests\Fixtures;
+namespace Symfony\Component\Serializer\Tests\Fixtures\Annotations;
 
 use Symfony\Component\Serializer\Annotation\Groups;
+use Symfony\Component\Serializer\Tests\Fixtures\GroupDummyInterface;
 
 /**
  * @author Kévin Dunglas <dunglas@gmail.com>