+++ src/Symfony/Component/Serializer/Tests/Fixtures/Attributes/GroupDummyChild.php
@@ -9,7 +9,7 @@
  * file that was distributed with this source code.
  */
 
-namespace Symfony\Component\Serializer\Tests\Fixtures;
+namespace Symfony\Component\Serializer\Tests\Fixtures\Attributes;
 
 class GroupDummyChild extends GroupDummy
 {