+++ src/Symfony/Component/Serializer/Tests/Mapping/Loader/AnnotationLoaderTest.php
@@ -11,23 +11,17 @@
 
 namespace Symfony\Component\Serializer\Tests\Mapping\Loader;
 
-use Doctrine\Common\Annotations\AnnotationReader;
 use PHPUnit\Framework\TestCase;
 use Symfony\Component\Serializer\Mapping\AttributeMetadata;
 use Symfony\Component\Serializer\Mapping\ClassDiscriminatorMapping;
 use Symfony\Component\Serializer\Mapping\ClassMetadata;
 use Symfony\Component\Serializer\Mapping\Loader\AnnotationLoader;
-use Symfony\Component\Serializer\Tests\Fixtures\AbstractDummy;
-use Symfony\Component\Serializer\Tests\Fixtures\AbstractDummyFirstChild;
-use Symfony\Component\Serializer\Tests\Fixtures\AbstractDummySecondChild;
-use Symfony\Component\Serializer\Tests\Fixtures\IgnoreDummy;
-use Symfony\Component\Serializer\Tests\Fixtures\AbstractDummyThirdChild;
 use Symfony\Component\Serializer\Tests\Mapping\TestClassMetadataFactory;
 
 /**
  * @author Kévin Dunglas <dunglas@gmail.com>
  */
-class AnnotationLoaderTest extends TestCase
+abstract class AnnotationLoaderTest extends TestCase
 {
     /**
      * @var AnnotationLoader
@@ -36,7 +30,7 @@ class AnnotationLoaderTest extends TestCase
 
     protected function setUp(): void
     {
-        $this->loader = new AnnotationLoader(new AnnotationReader());
+        $this->loader = $this->createLoader();
     }
 
     public function testInterface()
@@ -46,28 +40,28 @@ class AnnotationLoaderTest extends TestCase
 
     public function testLoadClassMetadataReturnsTrueIfSuccessful()
     {
-        $classMetadata = new ClassMetadata('Symfony\Component\Serializer\Tests\Fixtures\GroupDummy');
+        $classMetadata = new ClassMetadata($this->getNamespace().'\GroupDummy');
 
         $this->assertTrue($this->loader->loadClassMetadata($classMetadata));
     }
 
     public function testLoadGroups()
     {
-        $classMetadata = new ClassMetadata('Symfony\Component\Serializer\Tests\Fixtures\GroupDummy');
+        $classMetadata = new ClassMetadata($this->getNamespace().'\GroupDummy');
         $this->loader->loadClassMetadata($classMetadata);
 
-        $this->assertEquals(TestClassMetadataFactory::createClassMetadata(), $classMetadata);
+        $this->assertEquals(TestClassMetadataFactory::createClassMetadata($this->getNamespace()), $classMetadata);
     }
 
     public function testLoadDiscriminatorMap()
     {
-        $classMetadata = new ClassMetadata(AbstractDummy::class);
+        $classMetadata = new ClassMetadata($this->getNamespace().'\AbstractDummy');
         $this->loader->loadClassMetadata($classMetadata);
 
-        $expected = new ClassMetadata(AbstractDummy::class, new ClassDiscriminatorMapping('type', [
-            'first' => AbstractDummyFirstChild::class,
-            'second' => AbstractDummySecondChild::class,
-            'third' => AbstractDummyThirdChild::class,
+        $expected = new ClassMetadata($this->getNamespace().'\AbstractDummy', new ClassDiscriminatorMapping('type', [
+            'first' => $this->getNamespace().'\AbstractDummyFirstChild',
+            'second' => $this->getNamespace().'\AbstractDummySecondChild',
+            'third' => $this->getNamespace().'\AbstractDummyThirdChild',
         ]));
 
         $expected->addAttributeMetadata(new AttributeMetadata('foo'));
@@ -78,7 +72,7 @@ class AnnotationLoaderTest extends TestCase
 
     public function testLoadMaxDepth()
     {
-        $classMetadata = new ClassMetadata('Symfony\Component\Serializer\Tests\Fixtures\MaxDepthDummy');
+        $classMetadata = new ClassMetadata($this->getNamespace().'\MaxDepthDummy');
         $this->loader->loadClassMetadata($classMetadata);
 
         $attributesMetadata = $classMetadata->getAttributesMetadata();
@@ -88,7 +82,7 @@ class AnnotationLoaderTest extends TestCase
 
     public function testLoadSerializedName()
     {
-        $classMetadata = new ClassMetadata('Symfony\Component\Serializer\Tests\Fixtures\SerializedNameDummy');
+        $classMetadata = new ClassMetadata($this->getNamespace().'\SerializedNameDummy');
         $this->loader->loadClassMetadata($classMetadata);
 
         $attributesMetadata = $classMetadata->getAttributesMetadata();
@@ -98,24 +92,27 @@ class AnnotationLoaderTest extends TestCase
 
     public function testLoadClassMetadataAndMerge()
     {
-        $classMetadata = new ClassMetadata('Symfony\Component\Serializer\Tests\Fixtures\GroupDummy');
-        $parentClassMetadata = new ClassMetadata('Symfony\Component\Serializer\Tests\Fixtures\GroupDummyParent');
+        $classMetadata = new ClassMetadata($this->getNamespace().'\GroupDummy');
+        $parentClassMetadata = new ClassMetadata($this->getNamespace().'\GroupDummyParent');
 
         $this->loader->loadClassMetadata($parentClassMetadata);
         $classMetadata->merge($parentClassMetadata);
 
         $this->loader->loadClassMetadata($classMetadata);
 
-        $this->assertEquals(TestClassMetadataFactory::createClassMetadata(true), $classMetadata);
+        $this->assertEquals(TestClassMetadataFactory::createClassMetadata($this->getNamespace(), true), $classMetadata);
     }
 
     public function testLoadIgnore()
     {
-        $classMetadata = new ClassMetadata(IgnoreDummy::class);
+        $classMetadata = new ClassMetadata($this->getNamespace().'\IgnoreDummy');
         $this->loader->loadClassMetadata($classMetadata);
 
         $attributesMetadata = $classMetadata->getAttributesMetadata();
         $this->assertTrue($attributesMetadata['ignored1']->isIgnored());
         $this->assertTrue($attributesMetadata['ignored2']->isIgnored());
     }
+
+    abstract protected function createLoader(): AnnotationLoader;
+    abstract protected function getNamespace(): string;
 }